Moved loginWithTonomy to ExternalUser class
