+++ src/services/keymanager.ts
@@ -5,8 +5,8 @@ enum KeyManagerLevel {
     PIN = 'PIN',
     FINGERPRINT = 'FINGERPRINT',
     LOCAL = 'LOCAL',
-    BROWSERLOCALSTORAGE = 'BROWSERLOCALSTORAGE',
-    BROWSERSESSIONSTORAGE = 'BROWSERSESSIONSTORAGE',
+    BROWSER_LOCAL_STORAGE = 'BROWSER_LOCAL_STORAGE',
+    BROWSER_SESSION_STORAGE = 'BROWSER_SESSION_STORAGE',
 }
 
 // eslint-disable-next-line @typescript-eslint/no-namespace
